scheduler/player: calculate and use the correct time that a track needs to be scheduled for
Previously the scheduler was assuming that sched_get_next() would be
called at the time that the next track would start playing.
In reality, though, the player needs to call it much earlier
and schedule this track to start at some time in the future,
which creates a problem for the scheduler to find the appropriate
zone from which to pick the next track.

For example, we might call sched_get_next() at 09:50 to schedule a track
for 10:02, and at 10:00 the scheduler is supposed to start playing tracks
from another zone. Unless we somehow tell sched_get_next() that the
track we are scheduling is for a future time, it will pick a track
from the previous zone.

With this patch now we keep track of this future point in time,
in player->sched_unix_time, and we instruct sched_get_next()
to fetch the track that is to be played at that time instead of right now.